requirements.txt
langchain-google-genai==4.2.1
chromadb
python-dotenv==1.2.2
pandas==2.0.0
requests